Fix an issue with the alpha version of PhpSpec 4.0
When using --prefer-lowest with a minimum stability of alpha, the version
installed is 4.0.0-alpha, between that version and version 4.0.4, scalar
type-hints were added, causing a fatal error because the methods in the
extension class to not match the signature of the interface.

There's no way to access the version number from an extension, so
reflection is used to determine which implementation to use.

This extension was written specifically for a project I am working on,
and it is not possible to upgrade the PhpSpec version from 4.0.0-alpha
because of other dependencies.